Separate out preprocess to speed up preprocessing
